rest API

+++ fs-project-backend/src/main/java/com/example/fs_project_backend/dao/UserTypeRepository.java
@@ -3,5 +3,5 @@ package com.example.fs_project_backend.dao;
 import com.example.fs_project_backend.entity.UserType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserTypeRepository extends JpaRepository<UserType, Long> {
+public interface UserTypeRepository extends JpaRepository<UserType, Integer> {
 }
