+++ fs-project-backend/src/main/java/com/example/fs_project_backend/service/UserTypeService.java
@@ -10,7 +10,7 @@ public interface UserTypeService {
 
     UserType create(UserType userType);
 
-    UserType update(Long id, UserType updated);
+    UserType update(Integer id, UserType updated);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
